Move errors to their own file in storage pkg
