+++ lib/portlayer/storage/errors.go
@@ -47,3 +47,21 @@ func IsErrVolumeInUse(err error) bool {
 
 	return ok
 }
+
+// VolumeStoreNotFoundError : custom error type for when we fail to find a target volume store
+type VolumeStoreNotFoundError struct {
+	Msg string
+}
+
+func (e VolumeStoreNotFoundError) Error() string {
+	return e.Msg
+}
+
+// VolumeExistsError : custom error type for when a create operation targets and already occupied ID
+type VolumeExistsError struct {
+	Msg string
+}
+
+func (e VolumeExistsError) Error() string {
+	return e.Msg
+}
